Move scenario helper methods to base pydantic class
