+++ src/guidellm/benchmark/scenario.py
@@ -1,11 +1,8 @@
-import json
 from collections.abc import Iterable
 from pathlib import Path
-from typing import Annotated, Any, Literal, Optional, TypeVar, Union
+from typing import Annotated, Any, Literal, Optional, Union
 
-import yaml
 from datasets import Dataset, DatasetDict, IterableDataset, IterableDatasetDict
-from loguru import logger
 from pydantic import BeforeValidator, Field, NonNegativeInt, PositiveFloat, PositiveInt
 from transformers.tokenization_utils_base import (  # type: ignore[import]
     PreTrainedTokenizerBase,
@@ -35,33 +32,8 @@ def parse_float_list(value: Union[str, float, list[float]]) -> list[float]:
         ) from err
 
 
-T = TypeVar("T", bound="Scenario")
-
-
 class Scenario(StandardBaseModel):
     target: str
-
-    @classmethod
-    def get_default(cls: type[T], field: str) -> Any:
-        """Get default values for model fields"""
-        return cls.model_fields[field].default
-
-    @classmethod
-    def from_file(
-        cls: type[T], filename: Union[str, Path], overrides: Optional[dict] = None
-    ) -> T:
-        try:
-            with open(filename) as f:
-                if str(filename).endswith(".yaml") or str(filename).endswith(".yml"):
-                    data = yaml.safe_load(f)
-                else:  # Assume everything else is json
-                    data = json.load(f)
-        except (json.JSONDecodeError, yaml.YAMLError) as e:
-            logger.error("Failed to parse scenario")
-            raise e
-
-        data.update(overrides)
-        return cls.model_validate(data)
 
 
 class GenerativeTextScenario(Scenario):
